Add styles of the "Footer" main elements
Styles have been added to the project that
describe the appearance and behavior of the
main elements of the "Footer".

Were added:
---

* socials.scss — footer section containing
the logo and links to social networks;

* site-navigation.scss — section of links
to navigate the site;

* contacts.scss — section with contact
information about the company;

* copyright.scss — the lowest section
of the "Footer", inside which is a link
to the privacy policy, as well as a link
to the developer of this site.

Also some service and universal style files
have been added.