Add logic to limit the count of characters
A script, which describes the logic of limiting
the number of characters in the text "About сompany",
was added to the project.

If the count of characters in the text "About company"
exceeds the allowed limit, then the text is cut off to
the required count of characters and an ellipsis is
added to the end.