Supplement the logic of the Modal
New event handlers have been added to
the logic of the modal window:
---

* onEscPress() — event handler for
Closing the modal window when the
"Escape" key is pressed;

* onModalOverlayClick() — event handler
for Closing modal window when clicking
on the overlay;

Added settings to add these event handlers
at the right time, as well as to remove
them on time.

Also added a little logic for focusing
the cursor in the "Name" input field when
opening a modal window.