Fix the Footer markup
The markup of the section headings in the Footer
has been fixed: now the headings are not focused
until the Mobile version of the site, because
are not interactive elements.

Starting with the Mobile version of the site, text
heading blocks are replaced with interactive buttons,
so that they can have focus.